+++ TombEngine/Game/control/control.cpp
@@ -196,6 +196,7 @@ GameStatus GamePhase(bool insideMenu)
 	UpdateBlood();
 	UpdateBubbles();
 	UpdateDebris();
+	UpdateGunFlashes();
 	UpdateGunShells();
 	UpdateFootprints();
 	UpdateSplashes();